prove #545 is fixed, and split the test files my additions pushed over the ceiling
## #545 -- already fixed by the 1.21.0 merge, now proven

The reported failure: a watch with BelowPrice = 500 set in USD, then a poll
returning 480 EUR. At ~1.08 USD/EUR that is about 518 USD -- ABOVE the goal --
but a bare `480 <= 500` fired BelowGoal and told the user their target was met.

Current main already refuses that comparison: a currency mismatch sets
skipThresholdChecks, so BelowGoal and the last-minute signal are not evaluated
at all, BelowPrice and AlertDropAbs are cleared, and
AlertsClearedByCurrencyChange surfaces the loss so the user knows to re-set
them. The fix is refusal rather than conversion because no FX rate exists at
this layer.

TRVL.CURRENCY.BLIND.1 pins it, sabotage-verified: removing the threshold gate
fires BelowGoal on 480 EUR against a 500 USD target.

## Test files split back under the 800-line ceiling

My own additions pushed two files over it. Split by subject rather than
allowlisted -- the ceiling is a DoD gate and the allowlist is empty of source
files on purpose:

  - store_lifecycle_test.go 1062 -> 191, with every Migrate test moving to
    migrate_test.go, and migration's history/backup/concurrency half moving on
    to migrate_history_test.go (the first split still left 884).
  - check_currency_change_test.go 821 -> 764, with the new threshold test in
    currency_threshold_test.go. That file covers WHEN a currency change is
    detected; the new one covers what must not be compared once it has been.

Worth recording for next time: scripts/ci/check-file-size.sh only inspects
TRACKED files, so a newly created file reports clean until it is git-added. The
first split looked fine locally and would have failed CI.

Verification: internal/watch 308 passed, go vet clean, gofmt clean, staticcheck
clean, golangci-lint clean, all four scripts/ci hygiene checks pass.
